add service layer documentation and exception handling at user repo impl
