+++ SpringRedis/src/main/java/com/webonise/controller/UserController.java
@@ -42,4 +42,3 @@ public class UserController {
         return userService.findAll();
     }	
 }
-
